requirements_pi.txt
PyQt5
rpi_ws281x
adafruit-circuitpython-neopixel
